:truck: move docx generation service to domain
