+++ mcr-core/mcr_meeting/app/domain/report_templates/detailed_synthesis.md.jinja
